simplify code by remove pre 3.5 compat checks
