+++ aiofiles/base.py
@@ -1,7 +1,6 @@
 """Various base classes."""
 import asyncio
-
-from ._compat import PY_35
+from collections.abc import Coroutine
 
 
 class AsyncBase:
@@ -10,29 +9,23 @@ class AsyncBase:
         self._loop = loop
         self._executor = executor
 
-    if PY_35:
-        @asyncio.coroutine
-        def __aiter__(self):
-            """We are our own iterator."""
-            return self
 
-        @asyncio.coroutine
-        def __anext__(self):
-            """Simulate normal file iteration."""
-            line = yield from self.readline()
-            if line:
-                return line
-            else:
-                raise StopAsyncIteration
+    @asyncio.coroutine
+    def __aiter__(self):
+        """We are our own iterator."""
+        return self
 
-if PY_35:
-    from collections.abc import Coroutine
-    base = Coroutine
-else:
-    base = object
+    @asyncio.coroutine
+    def __anext__(self):
+        """Simulate normal file iteration."""
+        line = yield from self.readline()
+        if line:
+            return line
+        else:
+            raise StopAsyncIteration
 
 
-class _ContextManager(base):
+class _ContextManager(Coroutine):
     __slots__ = ('_coro', '_obj')
 
     def __init__(self, coro):
@@ -73,27 +66,25 @@ class _ContextManager(base):
         resp = yield from self._coro
         return resp
 
-    if PY_35:
-        def __await__(self):
-            resp = yield from self._coro
-            return resp
+    def __await__(self):
+        resp = yield from self._coro
+        return resp
 
-        @asyncio.coroutine
-        def __aenter__(self):
-            self._obj = yield from self._coro
-            return self._obj
+    @asyncio.coroutine
+    def __aenter__(self):
+        self._obj = yield from self._coro
+        return self._obj
 
-        @asyncio.coroutine
-        def __aexit__(self, exc_type, exc, tb):
-            self._obj.close()
-            self._obj = None
+    @asyncio.coroutine
+    def __aexit__(self, exc_type, exc, tb):
+        self._obj.close()
+        self._obj = None
 
 
 class AiofilesContextManager(_ContextManager):
     """An adjusted async context manager for aiofiles."""
-    if PY_35:
-        @asyncio.coroutine
-        def __aexit__(self, exc_type, exc_val, exc_tb):
-            yield from self._obj.close()
-            self._obj = None
 
+    @asyncio.coroutine
+    def __aexit__(self, exc_type, exc_val, exc_tb):
+        yield from self._obj.close()
+        self._obj = None
